requirements.txt
Flask==2.3.3
huggingface-hub==0.16.4
Pillow==10.0.1